# Conflicts:
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Resources/OperatorApp.cs.resx
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Resources/OperatorApp.de-DE.resx
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Resources/OperatorApp.hr.resx
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Resources/OperatorApp.it.resx
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Resources/OperatorApp.resx
#	VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Resources/OperatorApp.sl-SI.resx

+++ VerticalWarehouses/Installer/Ferretto.VW.Installer/Scripts/remove_current_version.ps1
@@ -5,9 +5,9 @@ $installer="$(Install:Root:Path)\$(Installer:DirName)";
 $telemetry="$(Install:Root:Path)\$(TS:DirName)";
 
 # remove current app dirs
-Remove-Item $as -Force -Recurse;
-Remove-Item $ppc -Force -Recurse;
-Remove-Item $installer -Force -Recurse
-Remove-Item $telemetry -Force -Recurse
+if(Test-Path $as -PathType Container) {Remove-Item $as -Force -Recurse};
+if(Test-Path $ppc -PathType Container) {Remove-Item $ppc -Force -Recurse};
+if(Test-Path $installer -PathType Container) {Remove-Item $installer -Force -Recurse};
+if(Test-Path $telemetry -PathType Container) {Remove-Item $telemetry -Force -Recurse};
 
-Remove-Item "E:\Database\Telemetry\*.*" -Force -Recurse
+if(Test-Path "E:\Database\Telemetry" -PathType Container) {Remove-Item "E:\Database\Telemetry\*.*" -Force -Recurse};
